+++ src/App.tsx
@@ -1,6 +1,6 @@
 import Layout from "./layout/Layout";
 import ExportButton from "./components/ExportButton";
-import Main from "./components/Main";
+import Main from "./components/Dashboard";
 
 import "./App.css";
